requirements/dev.txt
-r base.txt
django-debug-toolbar==1.2
